backend/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.32.0
python-dotenv==1.1.0
pydantic==2.11.6
baml-py==0.89.0
chromadb==1.0.12
sentence-transformers==4.1.0
termcolor==3.1.0
